Renamed shot-scraper storyboard command to shot-scraper video
